// api/reputation/model/comment-details-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Accommodation Service - Accommodation Owner APIs
 * Api Documentation for Accommodation Owner Users
 *
 * The version of the OpenAPI document: Accommodation Owner
 * Contact: amir.hossein.khalouei@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CommentSourceDetailsDto } from './comment-source-details-dto';

/**
 * 
 * @export
 * @interface CommentDetailsDto
 */
export interface CommentDetailsDto {
    /**
     * 
     * @type {string}
     * @memberof CommentDetailsDto
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CommentDetailsDto
     */
    date?: string;
    /**
     * 
     * @type {number}
     * @memberof CommentDetailsDto
     */
    rate?: number;
    /**
     * 
     * @type {string}
     * @memberof CommentDetailsDto
     */
    text?: string | null;
    /**
     * 
     * @type {CommentSourceDetailsDto}
     * @memberof CommentDetailsDto
     */
    commentSource?: CommentSourceDetailsDto;
}